server-rest-api/pyproject.toml
[tool.poetry]
name = "server-rest-api"
version = "0.1.0"
description = ""
authors = ["Arnab Mukherjee <am0072008@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
